[WEB-1710] combine share UX
