+++ app/pages/share/index.js
@@ -1,4 +1,3 @@
 export { default as AccessManagement } from './AccessManagement';
-export { default as InviteMember } from './InviteMember';
-export { default as InviteClinic } from './InviteClinic';
+export { default as ShareInvite } from './ShareInvite';
 export { default as PatientInvites } from './PatientInvites';
